Backport patch to allow for disabling activities overview hot corner.
This patch fixes the gsettings schema location that the enable-hot-corners option is installed to. This was changed in GNOME Shell 3.34.0 and while Tweak Tools includes this in 3.34 series, we're holding back our Tweak Tools due to its requirement for the mobile-centric library libhandy. So let's backport this patch to resolve T9032.